rename to close-widget-events
