+++ assets/js/src/core/modules/widget-manager/events/close-widget-events.ts
